[build] Fixed get_time/put_time on Linux.
